+++ src/leaflet/widgets/index.js
@@ -1,8 +1,8 @@
 /* Copyright© 2000 - 2018 SuperMap Software Co.Ltd. All rights reserved.
  * This program are made available under the terms of the Apache License, Version 2.0
  * which accompanies this distribution and is available at/r* http://www.apache.org/licenses/LICENSE-2.0.html.*/
-import {OpenFileView, openFileView} from './openfile/OpenFileView';
-import {OpenFileViewModel, openFileViewModel} from './openfile/OpenFileViewModel';
+import {OpenFileView, openFileView} from './openFile/OpenFileView';
+import {OpenFileViewModel, openFileViewModel} from './openFile/OpenFileViewModel';
 // import {POISearchView, poiSearchView} from './poisearch/POISearchView';
 import {DataFlowView, dataFlowView} from './dataflow/DataFlowView';